+++ CMakeLists.txt
@@ -254,6 +254,7 @@ set(TESTS
         db/corruption_test.cc
         db/cuckoo_table_db_test.cc
         db/db_iter_test.cc
+        db/db_log_iter_test.cc
         db/db_test.cc
         db/db_compaction_filter_test.cc
         db/db_dynamic_level_test.cc
